Moved "configured templates" into template loader.
When WordPress calls "template_include", the first step is to find IF any
plugins what their templates to be loaded.  That is the point where you
want to apply the filters.

If no configurations are returned, then bail out.  Why?

There are no plugins that want to load template(s). Therefore, there
is nothing for the Template Loader to do. Just bail.