Moved Events configuration loading tasks.
The bootstrap should be for loading up the plugin files
and getting it ready to run.  As this plugin grows, it
has a lot of configuration loading tasks.  At this point,
it is best to capture all of those tasks into one files
whose sole purpose is to....you guessed it....load the
configuration files.