Replaced hard-coded post type archive checkers.
Instead of using a hard-coded post type archive check, we
now use the current queried post type, i.e. from $wp_query.

This solution allows the Template Loader to be uncoupled to
the plugins.  Instead, it checks if this is a post type archive
request.  If yes, then it attempts to locate its template file
in either the theme or registered plugin.