Rearchitected Shortcode Module.
One of the things that would be very handy is to enqueue assets
for a shortcode.  In other words, if a web page processes a shortcode,
then enqueue the asset, if there is one.  In order to accomplish
this feature, we need to keep track of how many times a shortcode
is processed.

The easiest way to track this is by building a container.

Stick with me in this thought process.

The Shortcode Handler has become complex.  Using OOP, we can
hide away the complexity, only exposing the methods we need
externally.  With this redesign, we can store the number of times
a shortcode fires within a static property.

Using OOP, this process is much cleaner.