Added on-demand enqueuing.
This "on-demand" enqueuing reduces having scripts loaded into
the web page when there are no FAQs on the page.

If the shortcode is firing for the first time, then the
script is enqueued into the footer of the web page.

If the template is loaded, then it enqueues the script.