Improved change_title_placeholder_text().
1. Converted the conditional into a guard clause.
If the post type is not right, stop processing the callback and bail out.

2. Else, just return the text you want.
There's no need to assign the text to a variable and then return. Instead, it's
faster to just return it.

Why? It's less steps for PHP to process.  When assigning to a variable, here's what happens:

1. PHP creates updates the memory location with the new value.
2. Then when it's time to return, it has to look up the value before returning.