Restored dynamic view fixture file for integration tests.
Why?

The point of the tests is to exercise the code that builds the HTML and renders it out. In the events plugin, we see that there are dynamic PHP variables within the view file. We want to test that the code is properly pulling the post's meta data and then inserting it into the proper places within the rendered HTML view.

To make this work, here are the changes:

1. Load a single view file for all tests
2. Moved creating the post to the setUp() method - to reduce the redundant code
3. Add the post meta after creating the post - adds it into the database -- then the rest of the rendering code will exercise as normal during the integration tests.