Changed the integration test to validate custom fields config initialized.
Part of the purpose of this function is to merge the original custom fields'
configuration parameters with the default config parameters.

Why?

To ensure that each parameter exists in the custom fields' config.

In doing so, runtime configs can specify only the parameters that are
different from the default structure. Why? Less code to manage, test,
and maintain.